--- a/gzip_test.go
+++ b/gzip_test.go
@@ -3,11 +3,12 @@ package gziphandler
 import (
 	"bytes"
 	"compress/gzip"
-	"github.com/stretchr/testify/assert"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestParseEncodings(t *testing.T) {
@@ -36,10 +37,7 @@ func TestGzipHandler(t *testing.T) {
 	testBody := "aaabbbccc"
 
 	// This just exists to provide something for GzipHandler to wrap.
-	handler := GzipHandler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/plain")
-		io.WriteString(w, testBody)
-	}))
+	handler := newTestHandler(testBody)
 
 	// requests without accept-encoding are passed along as-is
 
@@ -63,10 +61,55 @@ func TestGzipHandler(t *testing.T) {
 	assert.Equal(t, gzipStr(testBody), res2.Body.Bytes())
 }
 
+// --------------------------------------------------------------------
+
+func BenchmarkGzipHandler_Serial(b *testing.B) {
+	req, _ := http.NewRequest("GET", "/whatever", nil)
+	req.Header.Set("Accept-Encoding", "gzip")
+	handler := newTestHandler("aaabbbccc")
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		runBenchmark(b, req, handler)
+	}
+}
+
+func BenchmarkGzipHandler_Parallel(b *testing.B) {
+	req, _ := http.NewRequest("GET", "/whatever", nil)
+	req.Header.Set("Accept-Encoding", "gzip")
+	handler := newTestHandler("aaabbbccc")
+
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			runBenchmark(b, req, handler)
+		}
+	})
+}
+
+// --------------------------------------------------------------------
+
 func gzipStr(s string) []byte {
 	var b bytes.Buffer
 	w := gzip.NewWriter(&b)
 	io.WriteString(w, s)
 	w.Close()
 	return b.Bytes()
+}
+
+func runBenchmark(b *testing.B, req *http.Request, handler http.Handler) {
+	res := httptest.NewRecorder()
+	handler.ServeHTTP(res, req)
+	if code := res.Code; code != 200 {
+		b.Fatalf("Expected 200 but got %d", code)
+	} else if blen := res.Body.Len(); blen != 33 {
+		b.Fatalf("Expected complete response body, but got %d bytes", blen)
+	}
+}
+
+func newTestHandler(body string) http.Handler {
+	return GzipHandler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "text/plain")
+		io.WriteString(w, body)
+	}))
 }