BEFORE:

```shell
$ go test ./... -test.bench=. -test.benchmem
PASS
BenchmarkGzipHandler_Serial     3000      513371 ns/op   1457632 B/op       53 allocs/op
BenchmarkGzipHandler_Parallel     3000      523055 ns/op   1457632 B/op       53 allocs/op
ok    github.com/nytimes/gziphandler  3.233s
```

AFTER:

```
$ go test ./... -test.bench=. -test.benchmem
PASS
BenchmarkGzipHandler_Serial     5000      256076 ns/op     57567 B/op       32 allocs/op
BenchmarkGzipHandler_Parallel     5000      251386 ns/op     47680 B/op       32 allocs/op
ok    github.com/nytimes/gziphandler  2.599s
```