@adammck 2015-08-17
I'm in favor of improving performance, but I suspect that in practice, the majority of the time spent would be in compressing the payloads rather than instantiating the writers. The test payloads are only 33 bytes, after all. Would you mind updating the benchmarks to use a (significantly) larger payload and see how that affects things?