@dim 2015-08-18
Hey, updated the benchmarks. Obviously the importance of this optimisation diminishes with the increasing compression effort, but to me it's still well worth it. 

In our [fork](https://github.com/bsm/gziphandler/blob/master/gzip.go#L33) we also allow to configure a lower compression level, which significantly reduces CPU/compression time and increases the performance delta.

```
BenchmarkGzipHandler_S2k      5000      330696 ns/op   1462784 B/op      106 allocs/op
BenchmarkGzipHandler_S20k     2000      971565 ns/op   1495872 B/op      121 allocs/op
BenchmarkGzipHandler_S100k       300     4091846 ns/op   1635680 B/op      195 allocs/op
BenchmarkGzipHandler_P2k      5000      337403 ns/op   1462784 B/op      106 allocs/op
BenchmarkGzipHandler_P20k     2000      976987 ns/op   1495872 B/op      121 allocs/op
BenchmarkGzipHandler_P100k       300     4103702 ns/op   1635681 B/op      195 allocs/op

BenchmarkGzipHandler_S2k      5000      228732 ns/op     17893 B/op       84 allocs/op
BenchmarkGzipHandler_S20k     2000      872040 ns/op    136766 B/op      100 allocs/op
BenchmarkGzipHandler_S100k       300     3955646 ns/op    191786 B/op      173 allocs/op
BenchmarkGzipHandler_P2k     10000      220198 ns/op     17602 B/op       84 allocs/op
BenchmarkGzipHandler_P20k     2000      856099 ns/op    136766 B/op      100 allocs/op
BenchmarkGzipHandler_P100k       300     4006694 ns/op    191788 B/op      173 allocs/op
```